# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Luka API",
    "version": "1.0.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "health",
        "summary": "El proceso vive",
        "responses": {
          "200": {
            "description": "Proceso activo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Crea la cuenta con correo y contraseña y guarda las tres aceptaciones",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cuenta creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Inicia sesión; la respuesta es la misma exista o no el correo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión iniciada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "summary": "Rota el token de refresco",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Par nuevo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Revoca los tokens de refresco de este dispositivo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Sesión cerrada"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "operationId": "readiness",
        "summary": "PostgreSQL y Redis responden",
        "responses": {
          "200": {
            "description": "Listo para recibir tráfico",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            }
          },
          "503": {
            "description": "Alguna dependencia no responde",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Problem": {
        "description": "Error con formato problem+json (RFC 9457)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Demasiados intentos de credenciales o demasiadas peticiones",
        "headers": {
          "Retry-After": {
            "description": "Segundos que hay que esperar antes de reintentar",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/RateLimitProblem"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "No se pueden comprobar los intentos de credenciales; reintentar más tarde",
        "headers": {
          "Retry-After": {
            "description": "Segundos que hay que esperar antes de reintentar",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/UnavailableProblem"
            }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["type", "title", "status", "code"],
        "additionalProperties": {}
      },
      "RateLimitProblem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number",
            "const": 429
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": ["too_many_attempts", "rate_limited"]
          }
        },
        "required": ["type", "title", "status", "code"],
        "additionalProperties": {}
      },
      "UnavailableProblem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "number",
            "const": 503
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "const": "temporarily_unavailable"
          }
        },
        "required": ["type", "title", "status", "code"],
        "additionalProperties": {}
      },
      "Health": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "ok"
          }
        },
        "required": ["status"],
        "additionalProperties": false
      },
      "Readiness": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok", "error"]
          },
          "checks": {
            "type": "object",
            "properties": {
              "postgres": {
                "type": "string",
                "enum": ["ok", "error"]
              },
              "redis": {
                "type": "string",
                "enum": ["ok", "error"]
              }
            },
            "required": ["postgres", "redis"],
            "additionalProperties": false
          }
        },
        "required": ["status", "checks"],
        "additionalProperties": false
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "maxLength": 256
          },
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "appVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "displayName": {
            "default": "",
            "type": "string",
            "maxLength": 80
          },
          "consents": {
            "type": "object",
            "properties": {
              "version": {
                "type": "string",
                "minLength": 1,
                "maxLength": 40
              },
              "terms": {
                "type": "boolean"
              },
              "privacy": {
                "type": "boolean"
              },
              "adult": {
                "type": "boolean"
              }
            },
            "required": ["version", "terms", "privacy", "adult"],
            "additionalProperties": false
          }
        },
        "required": [
          "email",
          "password",
          "deviceId",
          "appVersion",
          "userId",
          "displayName",
          "consents"
        ],
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "maxLength": 256
          },
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "appVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          }
        },
        "required": ["email", "password", "deviceId", "appVersion"],
        "additionalProperties": false
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": ["refreshToken"],
        "additionalProperties": false
      },
      "TokenPair": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": ["accessToken", "refreshToken", "expiresIn"],
        "additionalProperties": false
      }
    }
  }
}